requirements.txt
# Web 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# Telegram 机器人
aiogram==3.2.0

# 数据库
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
aiomysql==0.2.0
asyncpg==0.29.0

# 配置管理
pydantic-settings==2.1.0
python-dotenv==1.0.0

# 工具
aiofiles==23.2.1
